Rename metrics extension to smallrye-metrics
