+++ bom/pom.xml
@@ -83,7 +83,7 @@
 
          <dependency>
             <groupId>org.jboss.shamrock</groupId>
-            <artifactId>shamrock-metrics-deployment</artifactId>
+            <artifactId>shamrock-smallrye-metrics-deployment</artifactId>
             <version>${project.version}</version>
             <scope>provided</scope>
          </dependency>
