+++ build-parent/pom.xml
@@ -279,12 +279,12 @@
          </dependency>
          <dependency>
             <groupId>org.jboss.shamrock</groupId>
-            <artifactId>shamrock-metrics-deployment</artifactId>
+            <artifactId>shamrock-smallrye-metrics-deployment</artifactId>
             <version>${project.version}</version>
          </dependency>
          <dependency>
             <groupId>org.jboss.shamrock</groupId>
-            <artifactId>shamrock-metrics-runtime</artifactId>
+            <artifactId>shamrock-smallrye-metrics-runtime</artifactId>
             <version>${project.version}</version>
          </dependency>
          <dependency>
