+++ extensions/pom.xml
@@ -40,7 +40,7 @@
 
         <!-- Monitoring -->
         <module>smallrye-health</module>
-        <module>metrics</module>
+        <module>smallrye-metrics</module>
         <module>smallrye-opentracing</module>
         <module>fault-tolerance</module>
         <module>jaeger</module>
