+++ extensions/smallrye-metrics/deployment/pom.xml
@@ -19,15 +19,15 @@
          xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
          xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
     <parent>
-        <artifactId>shamrock-metrics</artifactId>
+        <artifactId>shamrock-smallrye-metrics</artifactId>
         <groupId>org.jboss.shamrock</groupId>
         <version>1.0.0.Alpha1-SNAPSHOT</version>
         <relativePath>../</relativePath>
     </parent>
     <modelVersion>4.0.0</modelVersion>
 
-    <artifactId>shamrock-metrics-deployment</artifactId>
-    <name>Shamrock - Metrics - Deployment</name>
+    <artifactId>shamrock-smallrye-metrics-deployment</artifactId>
+    <name>Shamrock - SmallRye Metrics - Deployment</name>
 
     <dependencies>
         <dependency>
@@ -44,7 +44,7 @@
         </dependency>
         <dependency>
             <groupId>org.jboss.shamrock</groupId>
-            <artifactId>shamrock-metrics-runtime</artifactId>
+            <artifactId>shamrock-smallrye-metrics-runtime</artifactId>
         </dependency>
     </dependencies>
