+++ extensions/smallrye-metrics/pom.xml
@@ -26,8 +26,8 @@
     </parent>
     <modelVersion>4.0.0</modelVersion>
 
-    <artifactId>shamrock-metrics</artifactId>
-    <name>Shamrock - Metrics</name>
+    <artifactId>shamrock-smallrye-metrics</artifactId>
+    <name>Shamrock - SmallRye Metrics</name>
     <packaging>pom</packaging>
     <modules>
         <module>deployment</module>
