+++ extensions/smallrye-metrics/runtime/src/main/java/org/jboss/shamrock/smallrye/metrics/runtime/LambdaCounter.java
@@ -14,30 +14,50 @@
  * limitations under the License.
  */
 
-package org.jboss.shamrock.metrics.runtime;
+package org.jboss.shamrock.smallrye.metrics.runtime;
 
 import java.util.concurrent.Callable;
 
-import org.eclipse.microprofile.metrics.Gauge;
+import org.eclipse.microprofile.metrics.Counter;
 
-/**
- * Created by bob on 7/31/18.
- */
-public class LambdaGauge implements Gauge {
+public class LambdaCounter implements Counter {
 
-    public LambdaGauge(Callable<Number> callable) {
+    private static final String MUST_NOT_BE_CALLED = "Must not be called";
+
+    public LambdaCounter(Callable<Long> callable) {
         this.callable = callable;
     }
 
     @Override
-    public Number getValue() {
+    public void inc() {
+        throw new IllegalStateException(MUST_NOT_BE_CALLED);
+    }
+
+    @Override
+    public void inc(long n) {
+        throw new IllegalStateException(MUST_NOT_BE_CALLED);
+    }
+
+    @Override
+    public void dec() {
+        throw new IllegalStateException(MUST_NOT_BE_CALLED);
+    }
+
+    @Override
+    public void dec(long n) {
+        throw new IllegalStateException(MUST_NOT_BE_CALLED);
+    }
+
+    @Override
+    public long getCount() {
         try {
             return this.callable.call();
         } catch (Exception e) {
             e.printStackTrace();
         }
+
         return 0;
     }
 
-    private final Callable<Number> callable;
+    private final Callable<Long> callable;
 }
