+++ integration-tests/main/pom.xml
@@ -50,7 +50,7 @@
         </dependency>
         <dependency>
             <groupId>org.jboss.shamrock</groupId>
-            <artifactId>shamrock-metrics-deployment</artifactId>
+            <artifactId>shamrock-smallrye-metrics-deployment</artifactId>
             <scope>provided</scope>
         </dependency>
         <dependency>
